backend/requirements.txt
# Queryon backend – RAG, DB, vectorstore, LLM/embedding clients
# Install: pip install -r backend/requirements.txt

typing_extensions>=4.0
tiktoken>=0.5

sqlalchemy[asyncio]>=2.0
asyncpg>=0.29

qdrant-client>=1.7

openai>=1.40
google-genai>=1.0

pypdf>=4.0
python-docx>=1.0
